optimize pytest exit
