+++ pytest/test_mesh_tags_simple.py
@@ -12,25 +12,6 @@ import numpy as np
 import PyOmega_h as omega_h
 
 
-# Create a global library instance to avoid Kokkos/MPI finalization issues
-# The library must outlive all mesh objects and is intentionally NOT deleted
-# to avoid MPI finalization order issues (see PyOmega_h_library.cpp)
-_global_lib = None
-
-
-@pytest.fixture(scope="session")
-def lib():
-    """Session-scoped library instance shared across all tests.
-    
-    Note: Cleanup is handled by conftest.py using os._exit() to avoid
-    MPI/Kokkos finalization order issues.
-    """
-    global _global_lib
-    if _global_lib is None:
-        _global_lib = omega_h.OmegaHLibrary()
-    return _global_lib
-
-
 @pytest.fixture
 def mesh(lib):
     """Create a simple 2D box mesh for testing."""
